# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Health Tracker \u2014 API Gateway",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/training/stats": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Training Stats",
        "operationId": "training_stats_api_v1_training_stats_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Default: 8 weeks ago.",
              "title": "From"
            },
            "description": "Default: 8 weeks ago."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Default: today.",
              "title": "To"
            },
            "description": "Default: today."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/training": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Create Training",
        "operationId": "create_training_api_v1_training_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingSessionIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "training"
        ],
        "summary": "List Training",
        "operationId": "list_training_api_v1_training_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "swim",
                    "gym"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by session type.",
              "title": "Type"
            },
            "description": "Filter by session type."
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingSessionOut"
                  },
                  "title": "Response List Training Api V1 Training Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/training/{session_id}": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Get Training",
        "operationId": "get_training_api_v1_training__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/training/{session_id}/sets": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Add Sets",
        "description": "Log sets against a session; free-text exercise names resolve to the catalog.",
        "operationId": "add_sets_api_v1_training__session_id__sets_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSetsIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Search Exercises",
        "operationId": "search_exercises_api_v1_exercises_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Name/alias substring for autocomplete.",
              "title": "Q"
            },
            "description": "Name/alias substring for autocomplete."
          },
          {
            "name": "muscle",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Muscle"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseOut"
                  },
                  "title": "Response Search Exercises Api V1 Exercises Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Create Exercise",
        "operationId": "create_exercise_api_v1_exercises_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddSetsIn": {
        "properties": {
          "sets": {
            "items": {
              "$ref": "#/components/schemas/TrainingSetIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Sets"
          }
        },
        "type": "object",
        "required": [
          "sets"
        ],
        "title": "AddSetsIn"
      },
...
      "ExerciseIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "push",
                  "pull",
                  "squat",
                  "hinge",
                  "carry",
                  "core",
                  "swim",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "primary_muscle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Muscle"
          },
          "equipment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment"
          },
          "is_unilateral": {
            "type": "boolean",
            "title": "Is Unilateral",
            "default": false
          },
          "is_bodyweight": {
            "type": "boolean",
            "title": "Is Bodyweight",
            "default": false
          },
          "aliases": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aliases"
          },
          "muscles": {
            "items": {
              "$ref": "#/components/schemas/ExerciseMuscleIn"
            },
            "type": "array",
            "title": "Muscles",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ExerciseIn"
      },
      "ExerciseMuscleIn": {
        "properties": {
          "muscle": {
            "type": "string",
            "title": "Muscle"
          },
          "role": {
            "type": "string",
            "enum": [
              "primary",
              "secondary"
            ],
            "title": "Role",
            "default": "primary"
          }
        },
        "type": "object",
        "required": [
          "muscle"
        ],
        "title": "ExerciseMuscleIn"
      },
      "ExerciseMuscleOut": {
        "properties": {
          "muscle": {
            "type": "string",
            "title": "Muscle"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "muscle",
          "role"
        ],
        "title": "ExerciseMuscleOut"
      },
      "ExerciseOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "primary_muscle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Muscle"
          },
          "equipment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment"
          },
          "is_unilateral": {
            "type": "boolean",
            "title": "Is Unilateral"
          },
          "is_bodyweight": {
            "type": "boolean",
            "title": "Is Bodyweight"
          },
          "aliases": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aliases"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "muscles": {
            "items": {
              "$ref": "#/components/schemas/ExerciseMuscleOut"
            },
            "type": "array",
            "title": "Muscles"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "category",
          "primary_muscle",
          "equipment",
          "is_unilateral",
          "is_bodyweight",
          "aliases",
          "is_active",
          "muscles"
        ],
        "title": "ExerciseOut"
      },
      "ExerciseStat": {
        "properties": {
          "exercise": {
            "type": "string",
            "title": "Exercise"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "sets": {
            "type": "integer",
            "title": "Sets"
          },
          "top_weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Weight Kg"
          },
          "best_e1rm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best E1Rm"
          },
          "best_e1rm_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best E1Rm Date"
          }
        },
        "type": "object",
        "required": [
          "exercise",
          "slug",
          "sets",
          "top_weight_kg",
          "best_e1rm",
          "best_e1rm_date"
        ],
        "title": "ExerciseStat"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MuscleVolume": {
        "properties": {
          "muscle": {
            "type": "string",
            "title": "Muscle"
          },
          "volume_load": {
            "type": "number",
            "title": "Volume Load"
          }
        },
        "type": "object",
        "required": [
          "muscle",
          "volume_load"
        ],
        "title": "MuscleVolume"
      },
...
      "TrainingSessionIn": {
        "properties": {
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts",
            "description": "Session start (UTC ISO; naive is assumed UTC)."
          },
          "type": {
            "type": "string",
            "enum": [
              "swim",
              "gym"
            ],
            "title": "Type"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          },
          "rpe": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rpe"
          },
          "load": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Load",
            "description": "Optional; auto-computed as duration_min \u00d7 rpe when omitted."
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "sets": {
            "items": {
              "$ref": "#/components/schemas/TrainingSetIn"
            },
            "type": "array",
            "title": "Sets",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "ts",
          "type"
        ],
        "title": "TrainingSessionIn"
      },
      "TrainingSessionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          },
          "rpe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rpe"
          },
          "load": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Load"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "kcal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kcal"
          },
          "avg_hr": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Hr"
          },
          "max_hr": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Hr"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "sets": {
            "items": {
              "$ref": "#/components/schemas/TrainingSetOut"
            },
            "type": "array",
            "title": "Sets"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ts",
          "type",
          "duration_min",
          "rpe",
          "load",
          "notes",
          "source",
          "kcal",
          "avg_hr",
          "max_hr",
          "distance_m",
          "sets"
        ],
        "title": "TrainingSessionOut"
      },
      "TrainingSetIn": {
        "properties": {
          "exercise": {
            "type": "string",
            "title": "Exercise",
            "description": "Free-text exercise / stroke, e.g. 'bench press', 'freestyle'."
          },
          "set_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set No"
          },
          "reps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reps"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "pace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pace"
          },
          "rpe": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rpe"
          },
          "is_warmup": {
            "type": "boolean",
            "title": "Is Warmup",
            "default": false
          },
          "added_weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Added Weight Kg"
          }
        },
        "type": "object",
        "required": [
          "exercise"
        ],
        "title": "TrainingSetIn"
      },
...
      "TrainingStats": {
        "properties": {
          "from": {
            "type": "string",
            "format": "date",
            "title": "From"
          },
          "to": {
            "type": "string",
            "format": "date",
            "title": "To"
          },
          "weekly_sets_per_muscle": {
            "items": {
              "$ref": "#/components/schemas/WeeklyMuscleSets"
            },
            "type": "array",
            "title": "Weekly Sets Per Muscle"
          },
          "volume_load_per_muscle": {
            "items": {
              "$ref": "#/components/schemas/MuscleVolume"
            },
            "type": "array",
            "title": "Volume Load Per Muscle"
          },
          "push_pull_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Push Pull Ratio"
          },
          "upper_lower_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upper Lower Ratio"
          },
          "per_exercise": {
            "items": {
              "$ref": "#/components/schemas/ExerciseStat"
            },
            "type": "array",
            "title": "Per Exercise"
          },
          "unresolved_exercises": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unresolved Exercises"
          }
        },
        "type": "object",
        "required": [
          "from",
          "to",
          "weekly_sets_per_muscle",
          "volume_load_per_muscle",
          "push_pull_ratio",
          "upper_lower_ratio",
          "per_exercise",
          "unresolved_exercises"
        ],
        "title": "TrainingStats"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeeklyMuscleSets": {
        "properties": {
          "week": {
            "type": "string",
            "format": "date",
            "title": "Week"
          },
          "muscle": {
            "type": "string",
            "title": "Muscle"
          },
          "sets": {
            "type": "number",
            "title": "Sets"
          }
        },
        "type": "object",
        "required": [
          "week",
          "muscle",
          "sets"
        ],
        "title": "WeeklyMuscleSets"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Shared gateway token (no 'Bearer ' prefix).",
        "scheme": "bearer"
      }
    }
  }
}